+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Added required `messaging_type` property (https://developers.facebook.com/docs/messenger-platform/send-messages#messaging_types) to all send operations.
 * Added support for the open graph template
 * Added support for additional webhook callbacks:
+  * Added support for policy enforcement
   * Added support for app roles
   * Added support for take thread control
   * Added support for pass thread control